requirements.txt
blinker==1.8.2
click==8.1.7
coverage==7.6.1
Flask==3.0.3
flask-swagger-ui==4.11.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
prometheus_client==0.20.0
prometheus_flask_exporter==0.23.1
python-dotenv==1.0.1
Werkzeug==3.0.4
